TabCompletion 		-- do not add greek letters to completion, but take them into 			    accoutn if they match (to prevent from destroying them), 			    in this case also add closing bracket. g:atp_imap_diacritics 	-- with g:atp_imap_diacritics_inteligent = 1 the work 			    even better (see doc). web_page		-- enhancments.

+++ ftplugin/ATP_files/compile.py
@@ -447,7 +447,8 @@ try:
             shutil.copy(file_cp, mainfile_dir)
     os.chdir(cwd)
     file_cp=basename+".aux"
-    shutil.copy(file_cp, os.path.join(mainfile_dir, basename+"._aux"))
+    if os.path.exists(os.path.join(mainfile_dir, basename+".aux")):
+        shutil.copy(file_cp, os.path.join(mainfile_dir, basename+"._aux"))
 
 ####################################
 #
